% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_metaphone.R
\name{double_metaphone}
\alias{double_metaphone}
\title{Double Metaphone}
\usage{
double_metaphone(word)
}
\arguments{
\item{word}{Character string}
}
\value{
Character vector of length 2 (primary, secondary)
}
\description{
Double Metaphone phonetic encoding algorithm
}
