% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmetaphone-package.R
\docType{package}
\name{fullmetaphone-package}
\alias{fullmetaphone}
\alias{fullmetaphone-package}
\title{fullmetaphone: Full-length Metaphone and Double Metaphone codes}
\description{
Phonetic codes give names that sound alike the same code, even when
they are spelt differently. This package provides
}
\details{
* [metaphone()], the original Metaphone algorithm (Philips 1990),
* [double_metaphone()], the improved Double Metaphone algorithm
  (Philips 2000), with a primary and a secondary code, and
* [sounds_like()], which compares two sets of names using either one.

Codes are returned in full. Most other implementations cut them to four
characters, which makes many different long names look the same. Use
`max_length = 4` when the traditional short codes are needed.
}
\section{Credits}{

Lawrence Philips designed both algorithms and published the original
code. The Double Metaphone rules here are translated from the C
implementation by Maurice Aubrey, which includes fixes by Kevin
Atkinson. The same C code is used in the 'PGRdup' package, whose
authors made Double Metaphone available in R and inspired this
package. The Metaphone rules and the reference test data come from
Apache Commons Codec. See the file `COPYRIGHTS` in the installed
package for details.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anilyayavar/fullmetaphone}
  \item Report bugs at \url{https://github.com/anilyayavar/fullmetaphone/issues}
}

}
\author{
\strong{Maintainer}: Anil Kumar Goyal \email{anilyayavar@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Anil Kumar Goyal \email{anilyayavar@gmail.com} [copyright holder]
  \item Atharv Tyagi \email{tatharv597@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Lawrence Philips (Designer of the Metaphone and Double Metaphone algorithms and author of their original code) [contributor]
  \item Maurice Aubrey (Author of the C implementation of Double Metaphone whose rules are followed here) [contributor]
  \item Kevin Atkinson (Bug fixes to that C implementation) [contributor]
  \item The Apache Software Foundation (Apache Commons Codec, the basis of metaphone() and of the test data) [copyright holder]
}

}
\keyword{internal}
