% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmetaphone-package.R
\docType{package}
\name{fullmetaphone-package}
\alias{fullmetaphone}
\alias{fullmetaphone-package}
\title{fullmetaphone: Full-length Metaphone and Double Metaphone codes}
\description{
Phonetic codes give names that sound alike the same code, even when
they are spelled differently. This package provides
}
\details{
* [metaphone()], the original Metaphone algorithm (Philips 1990),
* [double_metaphone()], the improved Double Metaphone algorithm
  (Philips 2000), with a primary and a secondary code,
* [sounds_like()], which compares two sets of names using either one,
  and
* [us_surnames], the 5000 most common surnames in the 2010 United
  States Census, for examples and testing.

Codes keep up to 32 characters by default, which is the complete code
for practically every real name. Most other implementations cut them
to four characters, which makes many different long names look the
same. All functions take a `max_length` argument to change this. Use
`max_length = Inf` for no limit, or `max_length = 4` for the
traditional short codes.

Start with `vignette("fullmetaphone")`, which explains the algorithms
and shows how to use the codes for deduplication, record linkage and
name search.
}
\section{Credits}{

Lawrence Philips designed both algorithms and published the original
code. The Double Metaphone rules here are translated from the C
implementation by Maurice Aubrey in the Perl module
\href{https://metacpan.org/pod/Text::DoubleMetaphone}{Text::DoubleMetaphone},
which includes fixes by Kevin Atkinson. The same C code is used in the
\href{https://CRAN.R-project.org/package=PGRdup}{'PGRdup'} package,
whose authors first made Double Metaphone available in R. The
Metaphone rules and the reference test data come from
\href{https://commons.apache.org/proper/commons-codec/}{Apache Commons Codec}.
See the file `COPYRIGHTS` in the installed package for details.
}

\references{
Philips, L. (1990). Hanging on the metaphone. *Computer Language*,
7(12), 38-43.

Philips, L. (2000). The double metaphone search algorithm. *C/C++
Users Journal*, 18(6), 38-43.
\url{https://web.archive.org/web/20250702064845/https://drdobbs.com/the-double-metaphone-search-algorithm/184401251}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anilyayavar/dblmetaphone}
  \item Report bugs at \url{https://github.com/anilyayavar/dblmetaphone/issues}
}

}
\author{
\strong{Maintainer}: Anil Kumar Goyal \email{anilyayavar@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Anil Kumar Goyal \email{anilyayavar@gmail.com} [copyright holder]
  \item Atharv Tyagi \email{tatharv597@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Lawrence Philips (Designer of the Metaphone and Double Metaphone algorithms and author of their original code) [contributor]
  \item Maurice Aubrey (Author of the C implementation of Double Metaphone whose rules are followed here) [contributor]
  \item Kevin Atkinson (Bug fixes to that C implementation) [contributor]
  \item The Apache Software Foundation (Apache Commons Codec, the basis of metaphone() and of the test data) [copyright holder]
}

}
\keyword{internal}
